+++ src/main/java/lotto/domain/Lottos.java
@@ -1,4 +1,6 @@
-package lotto;
+package lotto.domain;
+
+import lotto.enums.Rank;
 
 import java.util.ArrayList;
 import java.util.List;
