+++ src/test/java/lotto/LottoGeneratorTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.domain.LottoGenerator;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
