+++ src/test/java/lotto/LottoNumberTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.domain.LottoNumber;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
