+++ src/test/java/lotto/LottoPriceTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.domain.LottoPrice;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
