+++ src/test/java/lotto/LottoTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.domain.Lotto;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
